requirements.txt
Flask==3.1.0
SQLAlchemy==2.0.40
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
pytz==2025.2
python-dateutil==2.9.0.post0
celery==5.5.2
python-dotenv==1.1.0